Move `InReplyTo` to its own file.
