+++ libraries/matrix/api/src/main/kotlin/io/element/android/libraries/matrix/api/timeline/item/event/EventContent.kt
@@ -16,7 +16,6 @@
 
 package io.element.android.libraries.matrix.api.timeline.item.event
 
-import io.element.android.libraries.matrix.api.core.EventId
 import io.element.android.libraries.matrix.api.core.UserId
 import io.element.android.libraries.matrix.api.media.AudioInfo
 import io.element.android.libraries.matrix.api.media.FileInfo
@@ -35,33 +34,6 @@ data class MessageContent(
     val isThreaded: Boolean,
     val type: MessageType?
 ) : EventContent
-
-sealed interface InReplyTo {
-    /** The event details are not loaded yet. We can fetch them. */
-    data class NotLoaded(val eventId: EventId) : InReplyTo
-
-    /** The event details are pending to be fetched. We should **not** fetch them again. */
-    data object Pending : InReplyTo
-
-    /** The event details are available. */
-    data class Ready(
-        val eventId: EventId,
-        val content: EventContent,
-        val senderId: UserId,
-        val senderDisplayName: String?,
-        val senderAvatarUrl: String?,
-    ) : InReplyTo
-
-    /**
-     * Fetching the event details failed.
-     *
-     * We can try to fetch them again **with a proper retry strategy**, but not blindly:
-     *
-     * If the reason for the failure is consistent on the server, we'd enter a loop
-     * where we keep trying to fetch the same event.
-     * */
-    data object Error : InReplyTo
-}
 
 data object RedactedContent : EventContent
